refactor(frontend): extract api consumer component
